Widen the multisell weight check, which its own ingredient loop already does
MultiSellChoose totals up what the player is about to receive:

	int slots = 0;
	int weight = 0;
	...
	weight += e.getItemCount() * _amount * e.getWeight();

then hands the result to inv.validateWeight and inv.validateCapacity. Both of
those take a long. The range was being thrown away here, before the call, in
int arithmetic.

Sixty lines below, the ingredient loop guards exactly this hazard and gets it
right: ((long) e.getItemCount() * _amount) > Integer.MAX_VALUE, and it refuses
the exchange. The product loop above it does not, which is the asymmetry that
makes this a finding.

Measured before changing it. MultisellAmountLimit ships at 10000, and across the
74 shipped multisell files and their 7119 productions the worst case is
109,800,000 against an int ceiling of 2,147,483,647 -- nineteen times clear, and
nothing shipped overflows. But the threshold is count times weight above
214,748, and the heaviest item in the game weighs 10,980, so a custom production
of twenty of those crosses it. Multisell files are among the most commonly
edited on a private server. A weight that wraps negative passes the check it is
meant to fail, and the player walks away over the limit.

Both totals are long now and both multiplies are widened first.

Not changed, and measured for the same reason: the product count itself,
e.getItemCount() * _amount, is unguarded in the granting loop, but overflowing
it needs a production count above 214,748 and the largest in the shipped data is
700. Adding the symmetric guard there would be dead code.

+++ java/org/l2jmobius/gameserver/network/clientpackets/MultiSellChoose.java
@@ -117,8 +117,15 @@ public class MultiSellChoose extends ClientPacket
 				}
 				
 				final PlayerInventory inv = player.getInventory();
-				int slots = 0;
-				int weight = 0;
+				// long, and widened before each multiply. validateWeight and validateCapacity both
+				// take a long already, so the range was being thrown away here rather than at the
+				// call. The ingredient loop sixty lines below gets this right -- it guards with
+				// ((long) e.getItemCount() * _amount) > Integer.MAX_VALUE -- and the product loop
+				// did not. With MultisellAmountLimit at the shipped 10000, a production of twenty
+				// of the heaviest item in the game overflows, and a negative weight passes the
+				// check it is supposed to fail.
+				long slots = 0;
+				long weight = 0;
 				for (Ingredient e : entry.getProducts())
 				{
 					if (e.getItemId() < 0)
@@ -128,14 +135,14 @@ public class MultiSellChoose extends ClientPacket
 					
 					if (!e.isStackable())
 					{
-						slots += e.getItemCount() * _amount;
+						slots += (long) e.getItemCount() * _amount;
 					}
 					else if (player.getInventory().getItemByItemId(e.getItemId()) == null)
 					{
 						slots++;
 					}
 					
-					weight += e.getItemCount() * _amount * e.getWeight();
+					weight += (long) e.getItemCount() * _amount * e.getWeight();
 				}
 				
 				if (!inv.validateWeight(weight))
